requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
black==21.9b0
bleach==3.3.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
dataesg==0.0.2
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.2
entrypoints==0.3
filelock==3.0.12
huggingface-hub==0.0.18
idna==2.10
imageio==2.16.0
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.6.3
notebook==6.2.0
numexpr==2.7.2
numpy==1.20.1
oauthlib==3.1.0
orca==1.5.4
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.0
plotly==4.14.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.5.10
pytz==2021.1
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==21.0.1
regex==2021.9.30
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sacremoses==0.0.46
scikit-image==0.19.2
scipy==1.8.0
Send2Trash==1.5.0
sentencepiece==0.1.96
six==1.15.0
soupsieve==2.2
tables==3.6.1
TBB==0.1
terminado==0.9.2
testpath==0.4.4
tifffile==2022.2.9
tokenizers==0.10.3
toml==0.10.1
tomli==1.2.1
toolz==0.11.1
torch==1.10.2
torchvision==0.11.3
tornado==5.1.1
tqdm==4.62.3
traitlets==5.0.5
transformers==4.11.2
tweepy==3.10.0
typing-extensions==3.10.0.2
urllib3==1.26.2
virtualenv==20.7.2
vprof==0.38
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
